+++ include/command_macros_base.h
@@ -5,10 +5,10 @@
 
 #if IS_BIG_ENDIAN
 #if IS_64_BIT
-#define CMD_BBBB(a, b, c, d) ((uintptr_t)(_SHIFTL(a, 24, 8) | _SHIFTL(b, 16, 8) | _SHIFTL(c, 8, 8) | _SHIFTL(d, 0, 8)) << 32)
-#define CMD_BBH(a, b, c) ((uintptr_t)(_SHIFTL(a, 24, 8) | _SHIFTL(b, 16, 8) | _SHIFTL(c, 0, 16)) << 32)
-#define CMD_HH(a, b) ((uintptr_t)(_SHIFTL(a, 16, 16) | _SHIFTL(b, 0, 16)) << 32)
-#define CMD_W(a) ((uintptr_t)(a) << 32)
+#define CMD_BBBB(a, b, c, d) ((u32)(_SHIFTL(a, 24, 8) | _SHIFTL(b, 16, 8) | _SHIFTL(c, 8, 8) | _SHIFTL(d, 0, 8)) << 32)
+#define CMD_BBH(a, b, c) ((u32)(_SHIFTL(a, 24, 8) | _SHIFTL(b, 16, 8) | _SHIFTL(c, 0, 16)) << 32)
+#define CMD_HH(a, b) ((u32)(_SHIFTL(a, 16, 16) | _SHIFTL(b, 0, 16)) << 32)
+#define CMD_W(a) ((u32)(a) << 32)
 #else
 #define CMD_BBBB(a, b, c, d) (_SHIFTL(a, 24, 8) | _SHIFTL(b, 16, 8) | _SHIFTL(c, 8, 8) | _SHIFTL(d, 0, 8))
 #define CMD_BBH(a, b, c) (_SHIFTL(a, 24, 8) | _SHIFTL(b, 16, 8) | _SHIFTL(c, 0, 16))
@@ -21,7 +21,7 @@
 #define CMD_HH(a, b) (_SHIFTL(a, 0, 16) | _SHIFTL(b, 16, 16))
 #define CMD_W(a) (a)
 #endif
-#define CMD_PTR(a) ((uintptr_t)(a))
+#define CMD_PTR(a) ((u32)(a))
 
 #define CMD_HHHHHH(a, b, c, d, e, f) CMD_HH(a, b), CMD_HH(c, d), CMD_HH(e, f)
