+++ src/audio/copt/seq_channel_layer_process_script_copt.inc.c
@@ -96,11 +96,11 @@ struct Instrument **instOut = _instOut;\
         } \
         gi ## l:; \
     } \
-    if (((uintptr_t) gBankLoadedPool.persistent.pool.start <= (uintptr_t) inst \
-         && (uintptr_t) inst <= (uintptr_t)(gBankLoadedPool.persistent.pool.start \
+    if (((u32) gBankLoadedPool.persistent.pool.start <= (u32) inst \
+         && (u32) inst <= (u32)(gBankLoadedPool.persistent.pool.start \
                                           + gBankLoadedPool.persistent.pool.size)) \
-        || ((uintptr_t) gBankLoadedPool.temporary.pool.start <= (uintptr_t) inst \
-            && (uintptr_t) inst <= (uintptr_t)(gBankLoadedPool.temporary.pool.start \
+        || ((u32) gBankLoadedPool.temporary.pool.start <= (u32) inst \
+            && (u32) inst <= (u32)(gBankLoadedPool.temporary.pool.start \
                                              + gBankLoadedPool.temporary.pool.size))) { \
         (*adsr).envelope = (*inst).envelope; \
         (*adsr).releaseRate = (*inst).releaseRate; \
