+++ src/audio/seqplayer.c
@@ -1438,11 +1438,11 @@ u8 get_instrument(struct SequenceChannel *seqChannel, u8 instId, struct Instrume
         }
     }
 
-    if (((uintptr_t) gBankLoadedPool.persistent.pool.start <= (uintptr_t) inst
-         && (uintptr_t) inst <= (uintptr_t)(gBankLoadedPool.persistent.pool.start
+    if (((u32) gBankLoadedPool.persistent.pool.start <= (u32) inst
+         && (u32) inst <= (u32)(gBankLoadedPool.persistent.pool.start
                     + gBankLoadedPool.persistent.pool.size))
-        || ((uintptr_t) gBankLoadedPool.temporary.pool.start <= (uintptr_t) inst
-            && (uintptr_t) inst <= (uintptr_t)(gBankLoadedPool.temporary.pool.start
+        || ((u32) gBankLoadedPool.temporary.pool.start <= (u32) inst
+            && (u32) inst <= (u32)(gBankLoadedPool.temporary.pool.start
                                    + gBankLoadedPool.temporary.pool.size))) {
         adsr->envelope = inst->envelope;
         adsr->releaseRate = inst->releaseRate;
