+++ src/engine/graph_node_manager.c
@@ -7,11 +7,11 @@
 #if IS_64_BIT
 static s16 next_s16_in_geo_script(s16 **src) {
     s16 ret;
-    if (((uintptr_t)(*src) & 7) == 4) {
+    if (((u32)(*src) & 7) == 4) {
          *src += 2; // skip 32 bits
     }
     ret = *(*src)++;
-    if (((uintptr_t)(*src) & 7) == 4) {
+    if (((u32)(*src) & 7) == 4) {
          *src += 2; // skip 32 bits
     }
     return ret;
